Added DateModified field to Exercises.

+++ Code/CygSoft.SmartSession.Domain/Exercises/ExerciseService.cs
@@ -14,12 +14,21 @@ namespace CygSoft.SmartSession.Domain.Exercises
 
         public void Add(Exercise exercise)
         {
+            if (exercise.Id > 0)
+                throw new ArgumentException("A new exercise cannot have an id");
+
+            exercise.DateCreated = DateTime.Now;
+            exercise.DateModified = exercise.DateCreated;
+
             repository.Add(exercise);
             repository.SaveChanges();
         }
 
         public void Delete(int id)
         {
+            if (id <= 0)
+                throw new ArgumentException("The Id is invalid and must be greater than 0.");
+
             repository.Delete(id);
             repository.SaveChanges();
         }
@@ -36,6 +45,11 @@ namespace CygSoft.SmartSession.Domain.Exercises
 
         public void Update(Exercise exercise)
         {
+            if (exercise.Id <= 0)
+                throw new ArgumentException("An existing exercise must have an id");
+
+            exercise.DateModified = DateTime.Now;
+
             repository.Update(exercise);
             repository.SaveChanges();
         }
